// src/app/store/action-types/cart-action-types.ts
export enum CartActionTypes {
  GET_CART_DATA = '[Cart] Get cart data',
  GET_CART__DATA_SUCCESS = '[Cart] Get cart data success',
  GET_CART_DATA_FAILURE = '[Cart] Get cart data failure',

  DELETE_CART_ORDER = '[Cart] Delete cart order',
  DELETE_CART_ORDER_SUCCESS = '[Cart] Delete cart order success',
  DELETE_CART_ORDER_FAILURE = '[Cart] Delete cart order failure',

  UPDATE_CART_ORDER = '[Cart] Update cart order',
  UPDATE_CART_ORDER_SUCCESS = '[Cart] Update cart order success',
  UPDATE_CART_ORDER_FAILURE = '[Cart] Update cart order failure',

  ADD_CART_ORDER = '[Cart] Add cart order',
  ADD_CART_ORDER_SUCCESS = '[Cart] Add cart order success',
  ADD_CART_ORDER_FAILURE = '[Cart] Add cart order failure',
}